+++ ddoscoo-20200101/alibabacloud_ddoscoo20200101/models/_describe_domain_top_referer_request.py
@@ -10,6 +10,7 @@ class DescribeDomainTopRefererRequest(DaraModel):
         domain: str = None,
         end_time: int = None,
         limit: int = None,
+        query_type: str = None,
         region: str = None,
         start_time: int = None,
     ):
@@ -27,6 +28,7 @@ class DescribeDomainTopRefererRequest(DaraModel):
         # 
         # This parameter is required.
         self.limit = limit
+        self.query_type = query_type
         # The region in which your service is deployed. Valid values:
         # 
         # *   **cn**: a region in the Chinese mainland.
@@ -58,6 +60,9 @@ class DescribeDomainTopRefererRequest(DaraModel):
         if self.limit is not None:
             result['Limit'] = self.limit
 
+        if self.query_type is not None:
+            result['QueryType'] = self.query_type
+
         if self.region is not None:
             result['Region'] = self.region
 
@@ -76,6 +81,9 @@ class DescribeDomainTopRefererRequest(DaraModel):
 
         if m.get('Limit') is not None:
             self.limit = m.get('Limit')
+
+        if m.get('QueryType') is not None:
+            self.query_type = m.get('QueryType')
 
         if m.get('Region') is not None:
             self.region = m.get('Region')